+++ spec/models/early_career_payments/eligibility_answers_presenter_spec.rb
@@ -8,7 +8,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
       subject_to_disciplinary_action: false,
       pgitt_or_ugitt_course: :postgraduate,
       eligible_itt_subject: :chemistry,
-      teaching_subject_now: true
+      teaching_subject_now: true,
+      itt_academic_year: "2019_2020"
     }
   end
   let(:eligibility) { claim.eligibility }
@@ -21,7 +22,11 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
       [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
       [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "No", "supply-teacher"],
       [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"],
-      [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"],
+      [
+        I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"),
+        "Postgraduate",
+        "postgraduate-itt-or-undergraduate-itt-course"
+      ],
       [
         I18n.t("early_career_payments.questions.eligible_itt_subject", ug_or_pg: eligibility.pgitt_or_ugitt_course),
         "Chemistry",
@@ -31,6 +36,11 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         I18n.t("early_career_payments.questions.teaching_subject_now", eligible_itt_subject: eligibility.eligible_itt_subject),
         "Yes",
         "teaching-subject-now"
+      ],
+      [
+        I18n.t("early_career_payments.questions.itt_academic_year", start_or_complete: :start, ug_or_pg: eligibility.pgitt_or_ugitt_course),
+        "2019 - 2020",
+        "itt-year"
       ]
     ]
 
@@ -45,9 +55,10 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         has_entire_term_contract: true,
         employed_directly: true,
         subject_to_disciplinary_action: false,
-        pgitt_or_ugitt_course: :postgraduate,
+        pgitt_or_ugitt_course: :undergraduate,
         eligible_itt_subject: :modern_foreign_languages,
-        teaching_subject_now: true
+        teaching_subject_now: true,
+        itt_academic_year: "2018_2019"
       }
     end
 
@@ -60,7 +71,7 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"],
         [
           I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"),
-          "Postgraduate",
+          "Undergraduate",
           "postgraduate-itt-or-undergraduate-itt-course"
         ],
         [
@@ -72,6 +83,15 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
           I18n.t("early_career_payments.questions.teaching_subject_now", eligible_itt_subject: eligibility.eligible_itt_subject),
           "Yes",
           "teaching-subject-now"
+        ],
+        [
+          I18n.t(
+            "early_career_payments.questions.itt_academic_year",
+            start_or_complete: :complete,
+            ug_or_pg: eligibility.pgitt_or_ugitt_course
+          ),
+          "2018 - 2019",
+          "itt-year"
         ]
       ]